+++ Libplanet/Blockchain/Policies/BlockPolicy.cs
@@ -83,6 +83,8 @@ namespace Libplanet.Blockchain.Policies
         /// a <see cref="Block{T}"/> given the <see cref="Block{T}"/>'s index.
         /// Goes to <see cref="GetMaxTransactionsPerSignerPerBlock"/>.  Set to
         /// <see cref="GetMaxTransactionsPerBlock"/> by default.</param>
+        /// <param name="nativeTokens">A fixed set of <see cref="Currency"/> objects that are
+        /// supported by the blockchain as first-class citizens.  Empty by default.</param>
         public BlockPolicy(
             IAction? blockAction = null,
             TimeSpan? blockInterval = null,
@@ -97,7 +99,8 @@ namespace Libplanet.Blockchain.Policies
             Func<long, long>? getMaxBlockBytes = null,
             Func<long, int>? getMinTransactionsPerBlock = null,
             Func<long, int>? getMaxTransactionsPerBlock = null,
-            Func<long, int>? getMaxTransactionsPerSignerPerBlock = null)
+            Func<long, int>? getMaxTransactionsPerSignerPerBlock = null,
+            IImmutableSet<Currency>? nativeTokens = null)
         {
             BlockAction = blockAction;
             BlockInterval = blockInterval
@@ -107,6 +110,7 @@ namespace Libplanet.Blockchain.Policies
             MinimumDifficulty = minimumDifficulty
                 ?? DifficultyAdjustment<T>.DefaultMinimumDifficulty;
             CanonicalChainComparer = canonicalChainComparer ?? new TotalDifficultyComparer();
+            NativeTokens = nativeTokens ?? ImmutableHashSet<Currency>.Empty;
             _hashAlgorithmGetter = hashAlgorithmGetter ?? (_ => HashAlgorithmType.Of<SHA256>());
             _getMaxBlockBytes = getMaxBlockBytes ?? (_ => 100L * 1024L);
             _getMinTransactionsPerBlock = getMinTransactionsPerBlock ?? (_ => 0);
@@ -195,7 +199,7 @@ namespace Libplanet.Blockchain.Policies
 
         /// <inheritdoc cref="IBlockPolicy{T}.NativeTokens"/>
         // TODO: This should be configurable through the constructor.
-        public IImmutableSet<Currency> NativeTokens => ImmutableHashSet<Currency>.Empty;
+        public IImmutableSet<Currency> NativeTokens { get; }
 
         /// <summary>
         /// Targeted time interval between two consecutive <see cref="Block{T}"/>s.